requirements.txt
psycopg2
kafka-python
snowflake-connector-python
python-decouple